% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spDates.R
\name{calAll}
\alias{calAll}
\title{Calibrate all radiocarbon dates in a dataset, differentiating between
terrestrial and marine (shell) samples.}
\usage{
calAll(sites, c14bp, sd, material, curve = "intcal13")
}
\arguments{
\item{sites}{A dataframe with, minimally, the C14 ages, standard errors and
materials dated.}

\item{c14bp}{A string. Name of the field with the radiocarbon ages in C14
BP format.}

\item{sd}{A string. Name of the field with the standard errors of the C14
dates.}

\item{material}{A string. Name of the field with the materials dated.}

\item{curve}{A string. Calibration curve as per the rcarbon package.
Default is "intcal13".}
}
\value{
A dataframe with an added column "cal" with CalDates objects from
the rcarbon package.
}
\description{
Calibrate all radiocarbon dates in a dataset, differentiating between
terrestrial and marine (shell) samples.
}
