% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spDates.R
\docType{data}
\name{centers}
\alias{centers}
\title{Coordinates of 9 sites considered as potential centers of origin of the
Neolithic expansion. Modified from Pinhasi et al. (2005).}
\format{A data frame with 9 rows and 3 variables.
\itemize{
  \item Latitude. Site latitude in decimal degrees.
  \item Longitude. Site longitude in decimal degrees.
  \item Site. Site name.
}}
\usage{
centers
}
\description{
Coordinates of 9 sites considered as potential centers of origin of the
Neolithic expansion. Modified from Pinhasi et al. (2005).
}
\keyword{datasets}
