% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spDates.R
\docType{data}
\name{cost}
\alias{cost}
\title{Cost surface to calculate shortest paths (least-cost paths) of Neolithic
expansion from the Near East to Europe. Values are 1 (easiest, coast), 2
(land below 1750 m) and 3 (hardest, ocean and land above 1750 m).}
\format{A RasterLayer object.}
\usage{
cost
}
\description{
Cost surface to calculate shortest paths (least-cost paths) of Neolithic
expansion from the Near East to Europe. Values are 1 (easiest, coast), 2
(land below 1750 m) and 3 (hardest, ocean and land above 1750 m).
}
\keyword{datasets}
