% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spDates.R
\name{filterDates}
\alias{filterDates}
\title{Filter archaeological site coordinates and dates, retaining only the
earliest radiocarbon date per site.}
\usage{
filterDates(sites, c14bp)
}
\arguments{
\item{sites}{A SpatialPointsDataFrame object with archaeological sites and
associated radiocarbon ages.}

\item{c14bp}{A string. Name of the field with the radiocarbon ages in C14 BP
format.}
}
\value{
A SpatialPointsDataFrame object with the earliest C14 date for every
site.
}
\description{
Filter archaeological site coordinates and dates, retaining only the
earliest radiocarbon date per site.
}
