% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spDates.R
\name{interpolateIDW}
\alias{interpolateIDW}
\title{Interpolate using inverse distance weighting.}
\usage{
interpolateIDW(points, attr)
}
\arguments{
\item{points}{A SpatialPointsDataFrame object.}

\item{attr}{A string. Name of the field with values to interpolate.}
}
\value{
A RasterLayer.
}
\description{
Interpolate using inverse distance weighting.
}
