% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spDates.R
\name{iterateSites}
\alias{iterateSites}
\title{Perform regression of dates versus distances from multiple potential origins
in order to find the best model. It is also possible to specify multiple
distances for the spatial binning of the dates.}
\usage{
iterateSites(
  ftrSites,
  c14bp,
  siteNames,
  origins,
  binWidths = 0,
  nsim = 999,
  cost = NULL,
  method = "rma"
)
}
\arguments{
\item{ftrSites}{A SpatialPointsDataFrame object with associated earliest
C14 dates per site and respective calibrated distributions (CalDates
objects) in a field named "cal".}

\item{c14bp}{A string. Name of the field with the radiocarbon ages in C14
BP format.}

\item{siteNames}{A string. Name of the field with the site names or ids.}

\item{origins}{A SpatialPointsDataFrame object. The sites to be tested for
the most likely origin of expansion.}

\item{binWidths}{A number or vector of numbers. Width(s) of the spatial bins
in km.}

\item{nsim}{A number. Number of simulations to be run during the
bootstrapping procedure. Default is 999.}

\item{cost}{A RasterLayer with friction values to calculate least cost
path distances instead of great circle distances.}

\item{method}{A string. Method to be used in the regression. One of "rma"
or "ols". Default is "rma".}
}
\value{
a list with two elements, the result of the iteration over all
potential origins and the best model selected among those.
}
\description{
Perform regression of dates versus distances from multiple potential origins
in order to find the best model. It is also possible to specify multiple
distances for the spatial binning of the dates.
}
