% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spDates.R
\name{plot.dateMap}
\alias{plot.dateMap}
\title{Plot a map with the results of the iteration over multiple sites, showing
an interpolated surface of r values for the sites considered as potential
origins.}
\usage{
\method{plot}{dateMap}(dateMap)
}
\arguments{
\item{dateMap}{a dateMap object. One of the list elements returned from
the iterateSites() function.}
}
\value{
a spplot object.
}
\description{
Plot a map with the results of the iteration over multiple sites, showing
an interpolated surface of r values for the sites considered as potential
origins.
}
