% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spDates.R
\name{plot.dateModel}
\alias{plot.dateModel}
\title{Plot the results of the regression bootstrap on archaeological dates versus
distances from a hypothetical origin, with fitted lines.}
\usage{
plot.dateModel(dateModel)
}
\arguments{
\item{dateModel}{a dateModel object created with modelDates().}
}
\value{
a ggplot object.
}
\description{
Plot the results of the regression bootstrap on archaeological dates versus
distances from a hypothetical origin, with fitted lines.
}
