% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spDates.R
\name{sampleDates}
\alias{sampleDates}
\title{Extract a single year estimate for ranges of calibrated dates with a
probability given by the calibrated probability distribution.}
\usage{
sampleDates(calDates)
}
\arguments{
\item{calDates}{A CalDates object or a vector of CalDates.}
}
\value{
A vector of cal BP single year estimates.
}
\description{
Extract a single year estimate for ranges of calibrated dates with a
probability given by the calibrated probability distribution.
}
