% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spDates.R
\name{summary.dateModel}
\alias{summary.dateModel}
\title{Return summary statistics for a dateModel object.}
\usage{
\method{summary}{dateModel}(object, ...)
}
\arguments{
\item{object}{A dateModel object created with modelDates().}

\item{...}{ignored}
}
\value{
A dataframe with estimated start date and speed of advance.
}
\description{
Return summary statistics for a dateModel object.
}
